+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :feature:`-` Added new hook ``prepare_notification`` to process notifications before being sent by the notifications plugin
 * :feature:`662` Improve notifications plugin
 * :feature:`651` Add ``host_fqdn`` and ``host_name`` attributes to session
 * :feature:`647` Support internal plugins